refactor: rename camelCase files to kebab-case
Rename middleware, service, and route files to follow the project's
kebab-case file naming convention. No logic changes.
