+++ src/index.ts
@@ -1,7 +1,7 @@
 import { Hono } from 'hono'
 import { authMiddleware } from './middleware/auth'
-import { errorHandler } from './middleware/errorHandler'
-import { rateLimiterMiddleware } from './middleware/rateLimiter'
+import { errorHandler } from './middleware/error-handler'
+import { rateLimiterMiddleware } from './middleware/rate-limiter'
 import { cuisineRouter } from './routes/cuisines'
 import { ingredientRouter } from './routes/ingredients'
 import { tagRouter } from './routes/tags'
